fix: keep background agents' statuses live (fixes #3)
The pty-stream -> avatar-status parser lived in usePtyParser, which only
ran while an agent's profile panel was mounted. Any agent you were NOT
looking at kept whatever status it last showed: agents read as idle (or
stuck on working) until you clicked their profile. Worse, the idle-gated
inbox nudge in useHive skipped agents frozen on 'working', so their hive
mail sat unread and collaboration quietly stalled.

Move the parsing to a global watcher (ptyStatusWatcher.ts) that
subscribes to every live agent pty exactly once, app-wide, and reconciles
subscriptions against the agent list on store changes. usePtyParser
stays as a no-op shim so existing views keep compiling.

Hook events remain the authoritative status source (useHive effect #2);
the watcher is the refinement + fallback layer - including on Windows,
where the hook transport is currently broken (see #1 / PR #4).

`npm run typecheck` passes.
